Create Scrambler_program.jpynb
Working with python:
Make a word scrambler program to scramble every word in the file. Scrambling the word means "keeping first and last letter intact and scramble all other letters randomly". For example, if the following is written in the file: " Where is the abc.doc, xyz.html file?"  Then this will be changed to: "Werhe is the abd.coc, xzt.ymhl file? " Punctuation marks won't change their position.